Account Module Done

+++ app/src/main/java/com/heandroid/data/model/response/auth/forgot/password/SecurityCodeResponseModel.kt
@@ -4,4 +4,4 @@ import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class SecurityCodeResponseModel(var code: String?,var otpExpiryInSeconds: Int?,var referenceId: String?,var successful: Boolean) : Parcelable
+data class SecurityCodeResponseModel(var code: String?,var otpExpiryInSeconds: Long?,var referenceId: String?,var successful: Boolean) : Parcelable
